LPS-65332 Split item selector layout in private pages and public pages item selector
